API/requirements.txt
alembic==1.4.3
aniso8601==8.0.0
astroid==2.4.2
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.4
Flask==1.1.2
Flask-Jsonpify==1.5.0
Flask-Login==0.5.0
flask-marshmallow==0.14.0
Flask-Migrate==2.5.3
Flask-RESTful==0.3.8
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
idna==2.10
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.9.1
marshmallow-sqlalchemy==0.24.0
psycopg2==2.8.6
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.20
toml==0.10.1
urllib3==1.25.11
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3